+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.4.0
 + Remove `concat` implementation (the JDK has this)
++ Implement `suffle()` and `shuffle(RandomGenerator)`
 
 ### 0.3.0
 + Move minimum Java version from 22 to 23